Make tests run on CI
Specify the entrypoint and cmd on the tester_image2 so that it runs
tests. I was hoping to not have to duplicate the entrypoint and cmd
here, but the results from `docker inspect ${TESTER_IMAGE1_NAME} -f "{{ .Config.Cmd }}"`
were not suitable to pass back to `docker inspect` or `docker run`.