Add pytest-env to fix `pipenv run pytest` command
Previously tests would fail when `pipenv run pytest` was run because the
`FLASK_ENV` was not set. Rather than having to remember to run
`FLAKS_ENV=test pipenv run pytest`, we can use `pytest-env` to set the
env variable. The `D:` prefix tells `pytest-env` to not override an
existing value for the variable.