Use hidden input to track environment names
Since we're using ally.js for modals, when the modal is open, other
elements are disabled. Since the form is submitted when the "Create"
button is clicked and browsers do not submit disabled form fields, the
form was not submitting environment names when creating a project.

This workaround creates hidden inputs to track the environment names.
Since they're hidden, they are not disabled by ally.js, so the fields
are properly submitted.

This should be the only form affected, since this is the only form
submitted via a modal button.