Update uswds dep to fix sub-dep security vulnerability
There was a security vulnerability in `cached-path-relative`: https://hackerone.com/reports/390847

The package is a dependency of `browserify` which is a dependency of
`uswds`, which is one of our dependencies. Upgrading `uswds` to the
latest version installs the latest version of `cached-path-relative`
that has the security fix.

This update also requires bringing in a new static image that was
recently added in `uswds`.