Add vue component for confirmation popover
Have the confirmation popover in a separate Vue component fixes a bug in
IE that was causing the `form` element in the popover to be ignored.

Since `form`s cannot be nested, the `form` element in the popover was
being discarded by IE and the revoke/resend invitation buttons did
nothing.

Breaking the functionality into a Vue component moves the `form` into a
separate template. When the popover is displayed, the component is added
to the DOM at the end, so the `form` is properly not-nested.