Show errors on new task order form when fields are invalid
Form errors were not being shown because a new form instance was being
created each time the `form` property was being called. Since the
`validate` method on the form adds the errors to the form instance, this
was causing no errors to be shown even if the form were invalid.