Use `deepcopy` to copy form section info
Previously, copying the form sections was using `list.copy` which
creates a shallow copy of the list. Since the `TASK_ORDER_SECTIONS` list
contains a couple dictionaries, this was just creating another list with
references to the same dictionaries.

Therefore, when a section was marked as completed, it was updated
globally and visiting the new task order without filling anything out
would show some sections as completed.