Add config for EFS persistent volume in AWS k8s cluster.
Chose Elastic File Storage over EBS (Elastic Block Storage) because the
latter can only be mounted into a single node.

This adds the RBAC config and deployment for managing EFS mounts within
the cluster. Largely depends on this efs-provisioner config:
https://github.com/kubernetes-incubator/external-storage/tree/master/aws/efs
The config has been hard-copied into the repo and updated for future
reference. Note that the config requires an environment variable
substitution and cannot be applied directly to the cluster.