Specify Flask SERVER_NAME value for Celery worker.
The Celery worker cannot render URLs for the app without having a
SERVER_NAME value set. This currently breaks its ability to send
notifications when an environment is ready. This sets a default value
(localhost:8000) in the default config file and applies the correct
domains in the Kubernetes ConfigMaps for the AWS and Azure Celery
workers.