Adjust script to drop all tables and reset from scratch.
This updates the script for resetting the database so that it drops and
recreates all the tables, instead of disabling Postgres triggers and
truncating most of the tables. The latter strategy requires superuser
permissions in Postgres that the db user we manage in Azure does not
have. The script now:

- drops the tables
- reruns the alembic migrations
- reseeds the permission sets